add IT, unit tests, plugins for running Its and unit tests

+++ src/main/java/com/roman/models/Customer.java
@@ -15,7 +15,7 @@ public class Customer {
             strategy = GenerationType.SEQUENCE,
             generator = "customer_id_seq"
     )
-    private Long id;
+    private Integer id;
     @Column(
             nullable = false
     )
@@ -31,7 +31,7 @@ public class Customer {
 
     public Customer(){}
 
-    public Customer(Long id, String name, String email, Integer age) {
+    public Customer(Integer id, String name, String email, Integer age) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -45,11 +45,11 @@ public class Customer {
     }
 
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
