+++ src/main/java/com/roman/services/CustomerRowMapper.java
@@ -12,6 +12,6 @@ public class CustomerRowMapper implements RowMapper<Customer> {
 
     @Override
     public Customer mapRow(ResultSet rs, int rowNum) throws SQLException {
-        return new Customer(rs.getLong("id"), rs.getString("name"), rs.getString("email"), rs.getInt("age"));
+        return new Customer(rs.getInt("id"), rs.getString("name"), rs.getString("email"), rs.getInt("age"));
     }
 }
